requirements.txt
httpx
anthropic
google-auth
google-auth-oauthlib
google-api-python-client
beautifulsoup4
jinja2
python-dotenv
pytest
pytest-asyncio
python-jobspy
playwright
mcp
